Implement timeout for snippet extraction
Snippet extraction is done in a separate thread with limit set for the length of
the snippet processing.

Even though snippet extraction is done in a separate thread, only one snippet is
extracted at a time. Further improvement to extract multiple snippets in
multiple parallel threads (every with its own timeout) is necessary.

Useful resources:
https://stackoverflow.com/questions/20500003/setting-a-maximum-execution-time-for-a-method-thread
https://stackoverflow.com/questions/10504172/how-to-shutdown-an-executorservice
https://stackoverflow.com/questions/2733356/killing-thread-after-some-specified-time-limit-in-java
https://stackoverflow.com/questions/3590000/what-does-java-lang-thread-interrupt-do